+++ src/DependencyInjection/MonologExtension.php
@@ -197,8 +197,34 @@ final class MonologExtension extends Extension
                     ]);
                     $transport->setPublic(false);
 
+                    if (isset($handler['publisher']['encoder'])) {
+                        if ('compressed_json' === $handler['publisher']['encoder']) {
+                            $encoderClass = 'Gelf\Encoder\CompressedJsonEncoder';
+                        } elseif ('json' === $handler['publisher']['encoder']) {
+                            $encoderClass = 'Gelf\Encoder\JsonEncoder';
+                        } else {
+                            throw new \RuntimeException('The gelf message encoder must be either "compressed_json" or "json".');
+                        }
+
+                        $encoder = new Definition($encoderClass);
+                        $encoder->setPublic(false);
+
+                        $transport->addMethodCall('setMessageEncoder', [$encoder]);
+                    }
+
                     $publisher = new Definition('Gelf\Publisher', []);
                     $publisher->addMethodCall('addTransport', [$transport]);
+                    $publisher->setPublic(false);
+                } elseif (class_exists('Gelf\MessagePublisher')) {
+                    if (isset($handler['publisher']['encoder']) && 'compressed_json' !== $handler['publisher']['encoder']) {
+                        throw new \RuntimeException('The Gelf\MessagePublisher publisher supports only the compressed json encoding. Omit the option to use the default encoding or use "compressed_json" as the encoder option.');
+                    }
+                    $publisher = new Definition('Gelf\MessagePublisher', [
+                        $handler['publisher']['hostname'],
+                        $handler['publisher']['port'],
+                        $handler['publisher']['chunk_size'],
+                    ]);
+
                     $publisher->setPublic(false);
                 } else {
                     throw new \RuntimeException('The gelf handler requires the graylog2/gelf-php package to be installed.');